doublegenlist/requirements.txt
django-codenerix>=4.0.13
